Move steps for umask and file permissions to common
